Convert all compiler warnings to errors (-Werror)
Add -Werror flag to CMakeLists.txt to enforce warning-free builds
and fix all resulting sign-compare warnings:

**CMakeLists.txt changes:**
- Added -Werror to compile options for stricter code quality

**Fixed sign-compare warnings:**
- randread.cc: Changed ret to ssize_t, cast size to ssize_t in asserts
- randwrite.cc: Changed ret to ssize_t, cast size to ssize_t in asserts
- append.cc: Changed ret to ssize_t, cast size to ssize_t in assert
- salloc.cc: Cast set_capability to size_t in vector size comparisons

All warnings are now promoted to errors, ensuring code quality and
consistency. Both pxfs (19/19) and kvfs (6/7) tests pass with strict
compilation.

+++ libfs/bench/ubench/fs/append.cc
@@ -22,7 +22,7 @@ static int
 __ubench_fs_append(const char* root, int /*numops*/, int /*warmup_ops*/, size_t size)
 {
 	MEASURE_TIME_PREAMBLE
-	int                    ret = 0;
+	ssize_t                ret = 0;
 	unsigned long long     runtime;
 	hrtime_t               runtime_cycles = 0;
 	int                    fd;
@@ -52,7 +52,7 @@ __ubench_fs_append(const char* root, int /*numops*/, int /*warmup_ops*/, size_t 
 		assert(fd>0);
     	MEASURE_CYCLES_START
         	ret = fs_write(fd, buf, size);
-		assert(ret == size);
+		assert(ret == (ssize_t)size);
     	MEASURE_CYCLES_STOP
 		ADD_MEASURE_TIME_DIFF_CYCLES(runtime_cycles)
 		fs_close(fd);
